updated DeleteCardService

+++ app/services/cards/delete_card_service.rb
@@ -4,18 +4,19 @@ module Cards
   class DeleteCardService
     include Callable
 
-    attr_reader :user_id, :column_id, :id
+    attr_reader :current_user, :column_params, :id
 
-    def initialize(user_id, column_id, id)
-      @user_id = user_id
-      @column_id = column_id
+    def initialize(current_user, column_params, id)
+      @current_user = current_user
+      @column_params = column_params
       @id = id
     end
 
     def call
-      column = Column.find_by!(id: @column_id, user_id: @user_id)
+      raise Exceptions::Unauthorized, 'Unauthorized' unless @current_user.present?
+      column = Column.find_by!(id: @column_params[:column_id])
       card = column.cards.find_by!(id: @id)
-
+      raise Exceptions::Forbidden, 'Forbidden' unless card.user == @current_user
       card.destroy
 
       OpenStruct.new(success: card.destroyed?, errors: card.errors)
